ticketshow/requirements.txt
Flask==2.2
Flask-SQLAlchemy==3.0.3
Flask-Security
email_validator
bcrypt
flask-restful
Flask-JWT-Extended==4.4.4
flask-expects-json==1.7.0
Flask-WTF==1.0.1
redis==4.3.4
celery[redis]
WeasyPrint==52.5
Jinja2==3.1.2
beautifulsoup4
Flask-Caching
requests